Split agent and TrustedRouter parity gates
